% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_clean_source_medium.R
\name{ga_clean_source_medium}
\alias{ga_clean_source_medium}
\title{Cleans GA Data}
\usage{
ga_clean_source_medium(data, language = "es", remove_spam = FALSE)
}
\arguments{
\item{data}{a data frame from GA API. It must contain the column: ga:sourceMedium,
as the package works with this column to generate the ouputs.}

\item{language}{Choose a language for your sources column outputs.
Available languages: en, es, fr. More to add in the near future.}
}
\value{
The function returns the data frame with a new sources column with correct output ready to plot.
}
\description{
Cleans all source inputs from GA API.
}
\examples{
ga_clean_data(my_data, language="es")
}
