% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_get_pageviews.R
\name{ga_get_pageviews}
\alias{ga_get_pageviews}
\title{GA Monthly linechart graphic - sessions}
\usage{
ga_get_pageviews(view_id, start_date = start_date, final_date = final_date)
}
\arguments{
\item{data}{a data frame from GA API. It must contain the column: ga:sourceMedium,
as the package works with this column to generate the ouputs.}

\item{language}{Choose a language for your sources column outputs.
Available languages: en, es, fr. More to add in the near future.}
}
\value{
The function returns the data frame with a new sources column with correct output ready to plot.
}
\description{
Cleans all source inputs from GA API.
}
\examples{
ga_get_pageviews(view_id = 123456789, start_date = "2018-01-01", final_date = "2018-01-31")
}
