% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_get_sessions_by_campaign.R
\name{ga_get_sessions_by_campaign}
\alias{ga_get_sessions_by_campaign}
\title{GA Monthly linechart graphic - sessions}
\usage{
ga_get_sessions_by_campaign(view_id, start_date, final_date,
  device_category = F, language = "en")
}
\arguments{
\item{language}{Choose a language for your sources column outputs.
Available languages: en, es, fr. More to add in the near future.}

\item{data}{a data frame from GA API. It must contain the column: ga:sourceMedium,
as the package works with this column to generate the ouputs.}
}
\value{
The function returns the data frame with a new sources column with correct output ready to plot.
}
\description{
Cleans all source inputs from GA API.
}
\examples{
ga_clean_data(my_data, language="es")
}
