% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_pageviews_barchart_top_n.R
\name{ga_pageviews_barchart_top_n}
\alias{ga_pageviews_barchart_top_n}
\title{GA Monthly linechart graphic - sessions}
\usage{
ga_pageviews_barchart_top_n(data, top_n = 6, title = "Pageviews by source",
  x_title = "", y_title = "", subtitle = "", legend_title = "",
  label_size = 3)
}
\arguments{
\item{data}{a data frame from GA API. It must contain the column: ga:sourceMedium,
as the package works with this column to generate the ouputs.}

\item{language}{Choose a language for your sources column outputs.
Available languages: en, es, fr. More to add in the near future.}
}
\value{
Top pagevies by source. Will include more than n rows if there are ties.
}
\description{
Cleans all source inputs from GA API.
}
\examples{
ga_clean_data(my_data, language="es")
}
