% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gadata}
\alias{gadata}
\title{Real Google Analytics Data Sample}
\format{A data frame with 50 rows and 3 variables:
\describe{
  \item{date}{date of recorded data (2017-07-01)}
  \item{sourceMedium}{source and medium  (google / organic)}
  \item{sessions}{sessions (10, 50, 80, 100, 120)}
}}
\usage{
gadata
}
\description{
A dataset that contains real GA data for testing porpuses.
The variables are as follows:
}
\keyword{datasets}
