% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAnalyticsReports.R
\docType{package}
\name{googleAnalyticsReports}
\alias{googleAnalyticsReports}
\alias{googleAnalyticsReports-package}
\title{googleAnalyticsReports: A package for cleaning Google Analytics Data and plotting reports}
\description{
The googleAnalyticsReports package provides three categories of important functions:
foo, bar and baz.
}
\section{googleAnalyticsReports functions}{

ga_clean_data(data, language)
ga_sessions_per_month()
}

